Update patch set 2
Patch Set 2:

I strongly agree with Sylvian. Those issues should be addressed ASAP as they are they have serious security impact!

I'd like to use this opportunity to also emphasize that I'm really disappointed with the quality of code review in this project. Waste majority of patches is submitted and merged by the same company usually in less than an hour without any comment even if the code doesn't work (I found several cases like this while investigating one of OOM flags).

The code review process is *crucial* for the quality of ONAP code and security in a whole project. If you are unsure on the code you can always ask seccom for help to check if the code is good or bad. We are here to help thus I'd like to kindly ask you to be more conservative with merging patches (especially those which have a clear security impact) and put more stress on a proper code review phase.

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0